Move AboutType to types.ts
